Moved request to correct dir
