+++ app/Http/Controllers/AlbumController.php
@@ -27,6 +27,7 @@ use App\Http\Requests\Album\MergeAlbumsRequest;
 use App\Http\Requests\Album\MoveAlbumsRequest;
 use App\Http\Requests\Album\SetAlbumCoverRequest;
 use App\Http\Requests\Album\SetAlbumDescriptionRequest;
+use App\Http\Requests\Album\SetAlbumHeaderRequest;
 use App\Http\Requests\Album\SetAlbumLicenseRequest;
 use App\Http\Requests\Album\SetAlbumNSFWRequest;
 use App\Http\Requests\Album\SetAlbumProtectionPolicyRequest;
@@ -35,7 +36,6 @@ use App\Http\Requests\Album\SetAlbumsTitleRequest;
 use App\Http\Requests\Album\SetAlbumTagsRequest;
 use App\Http\Requests\Album\SetAlbumTrackRequest;
 use App\Http\Requests\Album\UnlockAlbumRequest;
-use App\Http\Requests\SetAlbumHeaderRequest;
 use App\Http\Resources\Collections\PositionDataResource;
 use App\Http\Resources\Models\AlbumResource;
 use App\Http\Resources\Models\SmartAlbumResource;
